+++ NEWS.md
@@ -10,8 +10,6 @@ New features
   for multiple game systems (#157).
 * Helper functions ``to_subpack`` and ``to_hexpack`` which given a piecepack configuration
   attempts to generate an appropriate matching (piecepack stackpack) subpack and hexpack (#161).
-* ``pp_cfg`` objects ``get_suit_color`` function is now vectorized 
-  (and by default now returns the suit colors for each suit).
 * New function ``file2grob`` that imports a given image file as a grob.
 
 Bug fixes and minor improvements
@@ -25,6 +23,8 @@ Bug fixes and minor improvements
   and shape/depth of pyramids.
 * Fixes bug in printing ``pp_cfg`` objects when a custom grob function had been set.
 * Fixes bug in setting individual suit colors with configurations list styles like ``suit_color.s2="white"``
+* ``pp_cfg`` objects' ``get_suit_color`` function is now vectorized 
+  (and by default now returns the suit colors for each suit).
 
 piecepackr 1.1.1
 ================